Switch int64 constructor to long long

+++ src/mongo/platform/decimal128.h
@@ -113,9 +113,9 @@ public:
     /**
      * This constructor is an interface for creating static constants
      */
-    Decimal128(const uint64_t dval[2]) : _value(dval) {}
+    Decimal128(const unsigned long long dval[2]) : _value(dval) {}
     Decimal128(int32_t int32Value);
-    Decimal128(int64_t int64Value);
+    Decimal128(long long int64Value);
 
     /**
      * This constructor takes a double and constructs a Decimal128 object
